Поделил код на модули по заданию #5
